+++ src/history.h
@@ -15,15 +15,15 @@ class history_buffers {
 	protected:
 		boost::circular_buffer< std::vector<double> > buffer;
 		double dt;
-		unsigned int n_vars;
+		unsigned long n_vars;
 	public:
-		history_buffers(int length, double dt, unsigned int n_vars);
+		history_buffers(unsigned long length, double dt, unsigned long n_vars);
 		history_buffers( const history_buffers& other );
-		int get_length();
-		virtual double get_value(double delay, int var_id)=0;
+		unsigned long get_length();
+		virtual double get_value(double delay, unsigned long var_id)=0;
 		local_state_type get_values(double delay);
-		double get_value_at(int position, int var_id);
-		local_state_type get_values_at(int position);
+		double get_value_at(unsigned long position, unsigned long var_id);
+		local_state_type get_values_at(unsigned long position);
 		void add_state( local_state_type state);
 		friend std::ostream& operator<< (std::ostream &out, history_buffers &h_bufs);
 };
@@ -33,22 +33,22 @@ class history_buffers {
 class history_factory
 {
 	public:
-		virtual history_buffers* create_history(int length, double dt, unsigned int n_vars)=0;
+		virtual history_buffers* create_history(unsigned long length, double dt, unsigned long n_vars)=0;
 };
 
 
 class lint_history: public history_buffers{
 	double linear_interpolate( double y1, double y2, double mu);
 	public:
-		lint_history(int length, double dt, unsigned int n_vars):history_buffers(length, dt, n_vars){};
+		lint_history(unsigned long length, double dt, unsigned long n_vars):history_buffers(length, dt, n_vars){};
 		lint_history( const lint_history& other ):history_buffers(other) { };
-		double get_value(double delay, int var_id);
+		double get_value(double delay, unsigned long var_id);
 };
 
 class lint_history_factory: public history_factory
 {
 	public:
-		lint_history* create_history(int length, double dt, unsigned int n_vars);
+		lint_history* create_history(unsigned long length, double dt, unsigned long n_vars);
 };
 
 #endif
