+++ src/mpi_integrator.h
@@ -25,9 +25,9 @@ class mpi_integrator: public integrator
 		neighbor_map_type send_ids;
 		// the actual integration scheme, performs one step and returns the
 		// length of the step in time
-		virtual double scheme(unsigned int node, local_state_type &new_state)=0; 
+		virtual double scheme(unsigned long node, local_state_type &new_state)=0; 
 		void step();
-		void dfun_eval(	unsigned int node,
+		void dfun_eval(	unsigned long node,
 						const local_state_type phi, 
 						double time_offset, 
 						local_state_type &dphidt);
@@ -42,14 +42,14 @@ class mpi_integrator: public integrator
 					const neighbor_map_type &send_node_ids);
 
 		~mpi_integrator();
-		void operator()(unsigned int n_steps);
+		void operator()(unsigned long n_steps);
 };
 
 // this duplicates euler_deterministic. Refactor with multiple inheritance?
 class mpi_euler_deterministic : public mpi_integrator
 {
 	private:
-		double scheme(unsigned int node, local_state_type &new_state);
+		double scheme(unsigned long node, local_state_type &new_state);
 	public:
 		mpi_euler_deterministic(	population_model *model,
 								population_coupling *coupling,
